Extract ContactListContent to its own file for clarity
