+++ examples/crm/src/contacts/ContactList.tsx
@@ -1,126 +1,37 @@
 /* eslint-disable import/no-anonymous-default-export */
 import * as React from 'react';
 import {
-    BulkActionsToolbar,
-    BulkDeleteButton,
     CreateButton,
     downloadCSV,
     ExportButton,
     List as RaList,
     Pagination,
-    RecordContextProvider,
-    ReferenceField,
-    SimpleListLoading,
     SortButton,
-    TextField,
     TopToolbar,
     useGetIdentity,
-    useListContext,
 } from 'react-admin';
 import type { Exporter } from 'react-admin';
-import {
-    List,
-    ListItem,
-    ListItemAvatar,
-    ListItemIcon,
-    ListItemSecondaryAction,
-    ListItemText,
-    Checkbox,
-    Typography,
-} from '@mui/material';
-import { Link } from 'react-router-dom';
-import { formatDistance } from 'date-fns';
 import jsonExport from 'jsonexport/dist';
 
-import { Avatar } from './Avatar';
-import { Status } from '../misc/Status';
-import { TagsList } from './TagsList';
 import { ContactListFilter } from './ContactListFilter';
+import { ContactListContent } from './ContactListContent';
 import { Contact, Company, Sale, Tag } from '../types';
 
-const ContactListContent = () => {
-    const {
-        data: contacts,
-        error,
-        isPending,
-        onToggleItem,
-        selectedIds,
-    } = useListContext<Contact>();
-    if (isPending) {
-        return <SimpleListLoading hasLeftAvatarOrIcon hasSecondaryText />;
-    }
-    if (error) {
-        return null;
-    }
-    const now = Date.now();
-
+export const ContactList = () => {
+    const { identity } = useGetIdentity();
+    if (!identity) return null;
     return (
-        <>
-            <BulkActionsToolbar>
-                <BulkDeleteButton />
-            </BulkActionsToolbar>
-            <List dense>
-                {contacts.map(contact => (
-                    <RecordContextProvider key={contact.id} value={contact}>
-                        <ListItem
-                            button
-                            component={Link}
-                            to={`/contacts/${contact.id}/show`}
-                        >
-                            <ListItemIcon sx={{ minWidth: '2.5em' }}>
-                                <Checkbox
-                                    edge="start"
-                                    checked={selectedIds.includes(contact.id)}
-                                    tabIndex={-1}
-                                    disableRipple
-                                    onClick={e => {
-                                        e.stopPropagation();
-                                        onToggleItem(contact.id);
-                                    }}
-                                />
-                            </ListItemIcon>
-                            <ListItemAvatar>
-                                <Avatar />
-                            </ListItemAvatar>
-                            <ListItemText
-                                primary={`${contact.first_name} ${contact.last_name}`}
-                                secondary={
-                                    <>
-                                        {contact.title} at{' '}
-                                        <ReferenceField
-                                            source="company_id"
-                                            reference="companies"
-                                            link={false}
-                                        >
-                                            <TextField source="name" />
-                                        </ReferenceField>
-                                        {contact.nb_tasks
-                                            ? ` - ${contact.nb_tasks} task${
-                                                  contact.nb_tasks > 1
-                                                      ? 's'
-                                                      : ''
-                                              }`
-                                            : ''}
-                                        &nbsp;&nbsp;
-                                        <TagsList />
-                                    </>
-                                }
-                            />
-                            <ListItemSecondaryAction>
-                                <Typography
-                                    variant="body2"
-                                    color="textSecondary"
-                                >
-                                    last activity{' '}
-                                    {formatDistance(contact.last_seen, now)} ago{' '}
-                                    <Status status={contact.status} />
-                                </Typography>
-                            </ListItemSecondaryAction>
-                        </ListItem>
-                    </RecordContextProvider>
-                ))}
-            </List>
-        </>
+        <RaList<Contact>
+            actions={<ContactListActions />}
+            aside={<ContactListFilter />}
+            perPage={25}
+            pagination={<Pagination rowsPerPageOptions={[10, 25, 50, 100]} />}
+            filterDefaultValues={{ sales_id: identity?.id }}
+            sort={{ field: 'last_seen', order: 'DESC' }}
+            exporter={exporter}
+        >
+            <ContactListContent />
+        </RaList>
     );
 };
 
@@ -156,21 +67,4 @@ const exporter: Exporter<Contact> = async (records, fetchRelatedRecords) => {
     return jsonExport(contacts, {}, (_err: any, csv: string) => {
         downloadCSV(csv, 'contacts');
     });
-};
-
-export const ContactList = () => {
-    const { identity } = useGetIdentity();
-    return identity ? (
-        <RaList<Contact>
-            actions={<ContactListActions />}
-            aside={<ContactListFilter />}
-            perPage={25}
-            pagination={<Pagination rowsPerPageOptions={[10, 25, 50, 100]} />}
-            filterDefaultValues={{ sales_id: identity?.id }}
-            sort={{ field: 'last_seen', order: 'DESC' }}
-            exporter={exporter}
-        >
-            <ContactListContent />
-        </RaList>
-    ) : null;
 };
